requirements.txt
#Use this command -> 'pip install -r requirements.txt' to install the dependencies automatically
getgauge==0.3.17
protobuf==3.20.0
selenium
websocket-client
requests
webdriver-manager